Create Code of Conduct
A primary goal for this project is to increase mobility and consistency for every individual's digital media profile. With that in mind, we have chosen a code of conduct to aid in guiding our future community. All activity related to this project should remain professional and unbiased. We value an inclusive community of users and contributors of all backgrounds and experience levels.